+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/wheeldrive/usecase/get_wheel_drives/GetWheelDrivesPort.java
@@ -1,4 +1,4 @@
-package softeer.be_my_car_master.api.wheeldrive.usecase.port;
+package softeer.be_my_car_master.api.wheeldrive.usecase.get_wheel_drives;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import softeer.be_my_car_master.domain.wheel_dirve.WheelDrive;
 import softeer.be_my_car_master.global.annotation.Port;
 
 @Port
-public interface WheelDrivePort {
+public interface GetWheelDrivesPort {
 
 	List<WheelDrive> findSelectableWheelDrivesByTrimId(Long trimId);
