+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/wheeldrive/usecase/get_wheel_drives/GetWheelDrivesUseCase.java
@@ -1,11 +1,10 @@
-package softeer.be_my_car_master.api.wheeldrive.usecase;
+package softeer.be_my_car_master.api.wheeldrive.usecase.get_wheel_drives;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 import lombok.RequiredArgsConstructor;
 import softeer.be_my_car_master.api.wheeldrive.dto.response.GetWheelDrivesResponse;
-import softeer.be_my_car_master.api.wheeldrive.usecase.port.WheelDrivePort;
 import softeer.be_my_car_master.domain.wheel_dirve.WheelDrive;
 import softeer.be_my_car_master.global.annotation.UseCase;
 
@@ -13,11 +12,11 @@ import softeer.be_my_car_master.global.annotation.UseCase;
 @RequiredArgsConstructor
 public class GetWheelDrivesUseCase {
 
-	private final WheelDrivePort wheelDrivePort;
+	private final GetWheelDrivesPort getWheelDrivePort;
 
 	public GetWheelDrivesResponse execute(Long trimId, Long engineId) {
-		List<WheelDrive> wheelDrives = wheelDrivePort.findSelectableWheelDrivesByTrimId(trimId);
-		List<Long> unselectableWheelDriveIds = wheelDrivePort.findUnselectableWheelDriveIdsByEngineId(engineId);
+		List<WheelDrive> wheelDrives = getWheelDrivePort.findSelectableWheelDrivesByTrimId(trimId);
+		List<Long> unselectableWheelDriveIds = getWheelDrivePort.findUnselectableWheelDriveIdsByEngineId(engineId);
 
 		List<WheelDrive> selectableWheelDrives = wheelDrives.stream()
 			.filter(wheelDrive -> wheelDrive.isSelectable(unselectableWheelDriveIds))
